Refactor: types 폴더 생성하여 video 타입 정리해서 사용
